% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dropNulls}
\alias{dropNulls}
\title{Drop NULL elements from a list}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{A list or vector}
}
\value{
The input with NULL elements removed
}
\description{
Removes NULL elements from a list or vector
}
