% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Left-hand side value}

\item{b}{Right-hand side value}
}
\value{
Returns \code{b} if \code{a} is NULL, otherwise returns \code{a}
}
\description{
Returns the right-hand side if the left-hand side is NULL
}
