% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeahead.R
\name{typeaheadInput}
\alias{typeaheadInput}
\title{Type-ahead text input}
\usage{
typeaheadInput(
  inputId,
  label = NULL,
  choices = character(),
  value = NULL,
  width = NULL,
  placeholder = NULL,
  items = 8,
  min_length = 1,
  hint = TRUE,
  options = list()
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{hint}{Logical; whether to show hints as you type}
}
\description{
Bootstrap-styled text box with client-side autocomplete.
}
