% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeahead.R
\name{updateTypeaheadInput}
\alias{updateTypeaheadInput}
\title{Update a typeahead input}
\usage{
updateTypeaheadInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  value = NULL
)
}
\arguments{
\item{session}{Shiny session object (default: getDefaultReactiveDomain()).}

\item{inputId}{Character string. The id of the typeahead input to update.}

\item{label}{Character string or NULL. New display label (optional).}

\item{choices}{Character vector or NULL. New choices (optional).}

\item{value}{Character string or NULL. New selected value (optional).}
}
\description{
Replace choices or value of typeaheadInput
}
